registration - confirm password and regex added

+++ src/pages/auth/Register.js
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import axios from "axios";
 import toast from "react-hot-toast";
 import { useNavigate } from "react-router-dom";
@@ -6,29 +6,61 @@ import { useNavigate } from "react-router-dom";
 import Jumbotron from "../../components/cards/Jumbotron";
 import { useAuth } from "../../context/auth";
 
+// const USER_REGEX = /^[a-zA-Z][a-zA-Z0-9-_]{3,23}$/;
+const PWD_REGEX = /^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%]).{8,24}$/;
+
 const Register = () => {
-  const [name, setName] = useState("krismat");
-  const [email, setEmail] = useState("kris1@gmail.com");
-  const [password, setPassword] = useState("123456");
+  const [name, setName] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [validPassword, setValidPassword] = useState(false);
+  const [matchPwd, setMatchPwd] = useState("");
+  const [validMatch, setValidMatch] = useState(false);
 
   const [auth, setAuth] = useAuth();
   const navigate = useNavigate();
 
+
+  
+  useEffect(() => {
+    const result = PWD_REGEX.test(password);
+    console.log(result);
+    console.log(password);
+    setValidPassword(result);
+    const match = password === matchPwd;
+    console.log(match)
+    setValidMatch(match);
+  }, [password, matchPwd]);
+
   const handleSubmit = async (e) => {
     e.preventDefault();
+
+    // const v1 = USER_REGEX.test(user);
+    const v2 = PWD_REGEX.test(password);
+
+    if (!v2) {
+      toast.error(` Password must have 6 to 24 characters.
+      Must begin with a letter.
+      Letters, numbers, underscores, hyphens allowed.`);
+      return;
+    }
+
+    if (validMatch !== true) {
+      toast.error(`Confirm your password`);
+      return;
+    }
+
     try {
-      const { data } = await axios.post(`/register`,
-        {
-          name,
-          email,
-          password,
-        }
-      );
+      const { data } = await axios.post(`/register`, {
+        name,
+        email,
+        password,
+      });
       console.log(data);
       if (data?.error) {
         toast.error(data.error);
       } else {
-        localStorage.setItem("auth", JSON.stringify(data))
+        localStorage.setItem("auth", JSON.stringify(data));
         setAuth({ ...auth, token: data.token, user: data.user });
         toast.success("Registration successful");
         navigate("/dashboard");
@@ -43,9 +75,11 @@ const Register = () => {
     <div>
       <Jumbotron title="Register" />
       <div className="container mt-5">
-        <div className="row">
-          <div className="col-md-6 offset-md-3">
-            <form onSubmit={handleSubmit}>
+        <div className="row justify-content-center">
+          <div className="col-md-3 ">
+            <form
+              onSubmit={handleSubmit}
+            >
               <input
                 type="text"
                 className="form-control mb-4 p-2"
@@ -67,7 +101,13 @@ const Register = () => {
                 placeholder="Enter your password"
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
-                autoFocus
+              />
+              <input
+                type="password"
+                className="form-control mb-4 p-2"
+                placeholder="Confirm your password"
+                value={matchPwd}
+                onChange={(e) => setMatchPwd(e.target.value)}
               />
 
               <button className="btn btn-primary" type="submit">
